move useGetSwapInfo to hooks hook + fetch swap info from RepayForm
