+++ src/pages/Swap/SubmitSection/index.tsx
@@ -7,11 +7,11 @@ import { useTranslation } from 'translation';
 import { Swap } from 'types';
 import { convertWeiToTokens, getContractAddress } from 'utilities';
 
+import { SwapError } from '../../../hooks/useGetSwapInfo';
 import { useAuth } from 'context/AuthContext';
 import useTokenApproval from 'hooks/useTokenApproval';
 
 import { FormError, FormValues } from '../types';
-import { SwapError } from '../useGetSwapInfo';
 import { useStyles } from './styles';
 
 const pancakeRouterContractAddress = getContractAddress('pancakeRouter');
